# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_IUser.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCredentials": {
				"$ref": "#/components/schemas/Pick_IUser.email-or-password_"
			},
			"INote": {
				"properties": {
					"title": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"createdBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_INote.title-or-content_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NoteCreateRequest": {
				"$ref": "#/components/schemas/Pick_INote.title-or-content_"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "jwt",
				"name": "jwt_token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "my-notes",
		"version": "1.0.0",
		"description": "my-notes exposes CRUD APIs required for notes management",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jay"
		}
	},
	"paths": {
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "hello@hello.com",
											"password": "don't tell anybody"
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "User already exists"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCredentials"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login success",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "hello@hello.com",
											"password": "don't tell anybody"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Invalid credentials"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCredentials"
							}
						}
					}
				}
			}
		},
		"/notes": {
			"get": {
				"operationId": "GetUserNotes",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/INote"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Invalid token"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNote",
				"responses": {
					"201": {
						"description": "Note created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/INote"
								},
								"examples": {
									"Example 1": {
										"value": {
											"title": "My first note",
											"content": "It's a bright day today"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Invalid token"
					},
					"409": {
						"description": "Note with same title exists"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NoteCreateRequest"
							}
						}
					}
				}
			}
		},
		"/notes/{noteId}": {
			"get": {
				"operationId": "GetUserNoteById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/INote"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Invalid token"
					},
					"404": {
						"description": "Note doesn't exists"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "noteId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}